backend/requirements.txt
# BizSakhi Backend Requirements
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0

# Database dependencies
sqlalchemy==2.0.23
aiosqlite==0.19.0
supabase==2.0.0
PyJWT>=2.8.0

# AI and ML dependencies
google-generativeai==0.3.2
anthropic>=0.7.0
groq>=0.4.0

# Speech and Audio processing
faster-whisper==0.10.0
gtts==2.4.0
pydub==0.25.1

# Image processing and OCR
Pillow==10.1.0
pytesseract==0.3.10
easyocr>=1.6.0
opencv-python>=4.8.0

# HTTP requests
requests>=2.31.0

# Additional utilities
numpy>=1.24.0
python-dateutil>=2.8.0

# Loan RAG Processor dependencies
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0